Add strong typing to separate snapshotter and image service opts
- Refactor common opts into pkg/nix/nix.go
- Moved pkg/nix/nix.go to pkg/nix/snapshotter.go
- Moved snapshotter tests to pkg/nix/snapshotter_* tests appropriately
